requirements.txt
pypdf==6.7.0
openai==2.21.0
qdrant-client==1.16.2
langchain-text-splitters==1.1.0
openai-whisper==20250625
torch==2.10.0
numpy==2.2.2
certifi==2024.12.14
